feat: add JWT authentication to review endpoints

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,4 +1,5 @@
 from flask_restx import Namespace, Resource, fields
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.services import facade
 
 api = Namespace('reviews', description='Review operations')
@@ -16,10 +17,19 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()
         review_data = api.payload
+        place = facade.get_place(review_data['place_id'])
 
+        if place.owner.id == current_user:
+            return {'error': 'You cannot review your own place.'}, 400
+
+        for review in facade.get_reviews_by_place(review_data['place_id']):
+            if current_user == review.user.id:
+                return {'error': 'You have already reviewed this place.'}, 400
         try:
             new_review = facade.create_review(review_data)
         except ValueError as e:
@@ -67,23 +77,33 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         review = facade.get_review(review_id)
+        current_user = get_jwt_identity()
         if not review:
             return {'error': 'Review not found'}, 404
-        
+
+        if current_user != review.user.id:
+            return {'error': 'Unauthorized action.'}, 403
+
         review_data = api.payload
         facade.update_review(review_id, review_data)
         return {"message": "Review updated successfully"}, 200
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
         review = facade.get_review(review_id)
+        current_user = get_jwt_identity()
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if current_user != review.user.id:
+            return {'error': 'Unauthorized action.'}, 403
 
         facade.delete_review(review_id)
         return {"message": "Review deleted successfully"}, 200
